+++ tokio-util/src/sync/mpsc.rs
@@ -1,36 +1,62 @@
 use futures_core::ready;
 use futures_sink::Sink;
+use tokio::sync::mpsc::OwnedPermit;
+use std::{fmt, mem};
 use std::pin::Pin;
-use std::sync::Arc;
 use std::task::{Context, Poll};
-use tokio::sync::mpsc::{error::SendError, Sender};
+use tokio::sync::mpsc::Sender;
 
 use super::ReusableBoxFuture;
 
-// This implementation was chosen over something based on permits because to get a
-// `tokio::sync::mpsc::Permit` out of the `inner` future, you must transmute the
-// lifetime on the permit to `'static`.
+/// Error returned by the `PollSender` when the channel is closed.
+#[derive(Debug)]
+pub struct PollSendError<T>(Option<T>);
+
+impl<T> PollSendError<T> {
+    /// Consumes the stored value, if any.
+    /// 
+    /// If this error was encountered when calling `start_send`, this will be the item that the
+    /// caller was attempting to send.  Otherwise, it will be `None`.
+    pub fn into_inner(self) -> Option<T> {
+        self.0
+    }
+}
+
+impl<T> fmt::Display for PollSendError<T> {
+    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(fmt, "channel closed")
+    }
+}
+
+impl<T: fmt::Debug> std::error::Error for PollSendError<T> {}
+
+#[derive(Debug)]
+enum State<T> {
+    Idle(Sender<T>),
+    Acquiring,
+    ReadyToSend(OwnedPermit<T>),
+    Closed,
+}
 
 /// A wrapper around [`mpsc::Sender`] that can be polled.
 ///
 /// [`mpsc::Sender`]: tokio::sync::mpsc::Sender
 #[derive(Debug)]
 pub struct PollSender<T> {
-    /// is none if closed
-    sender: Option<Arc<Sender<T>>>,
-    is_sending: bool,
-    inner: ReusableBoxFuture<Result<(), SendError<T>>>,
+    sender: Option<Sender<T>>,
+    state: State<T>,
+    acquire: ReusableBoxFuture<Result<OwnedPermit<T>, PollSendError<T>>>,
 }
 
-// By reusing the same async fn for both Some and None, we make sure every
-// future passed to ReusableBoxFuture has the same underlying type, and hence
-// the same size and alignment.
-async fn make_future<T>(data: Option<(Arc<Sender<T>>, T)>) -> Result<(), SendError<T>> {
+// Creates a future for acquiring a permit from the underlying channel.  This is used to ensure
+// there's capacity for a send to complete.
+//
+// By reusing the same async fn for both `Some` and `None`, we make sure every future passed to
+// ReusableBoxFuture has the same underlying type, and hence the same size and alignment.
+async fn make_acquire_future<T>(data: Option<Sender<T>>) -> Result<OwnedPermit<T>, PollSendError<T>> {
     match data {
-        Some((sender, value)) => sender.send(value).await,
-        None => unreachable!(
-            "This future should not be pollable, as is_sending should be set to false."
-        ),
+        Some(sender) => sender.reserve_owned().await.map_err(|_| PollSendError(None)),
+        None => unreachable!("this future should not be pollable in this state"),
     }
 }
 
@@ -38,183 +64,200 @@ impl<T: Send + 'static> PollSender<T> {
     /// Create a new `PollSender`.
     pub fn new(sender: Sender<T>) -> Self {
         Self {
-            sender: Some(Arc::new(sender)),
-            is_sending: false,
-            inner: ReusableBoxFuture::new(make_future(None)),
+            sender: Some(sender.clone()),
+            state: State::Idle(sender),
+            acquire: ReusableBoxFuture::new(make_acquire_future(None)),
         }
     }
 
-    /// Start sending a new item.
-    ///
-    /// This method panics if a send is currently in progress. To ensure that no
-    /// send is in progress, call `poll_send_done` first until it returns
-    /// `Poll::Ready`.
-    ///
-    /// If this method returns an error, that indicates that the channel is
-    /// closed. Note that this method is not guaranteed to return an error if
-    /// the channel is closed, but in that case the error would be reported by
-    /// the first call to `poll_send_done`.
-    pub fn start_send(&mut self, value: T) -> Result<(), SendError<T>> {
-        if self.is_sending {
-            panic!("start_send called while not ready.");
-        }
-        match self.sender.clone() {
-            Some(sender) => {
-                self.inner.set(make_future(Some((sender, value))));
-                self.is_sending = true;
-                Ok(())
-            }
-            None => Err(SendError(value)),
-        }
+    fn take_state(&mut self) -> State<T> {
+        mem::replace(&mut self.state, State::Closed)
     }
 
-    /// If a send is in progress, poll for its completion. If no send is in progress,
-    /// this method returns `Poll::Ready(Ok(()))`.
+    /// Attempts to prepare the sender to receive a value.
     ///
-    /// This method can return the following values:
+    /// This method must be called and return `Poll::Ready(Ok(()))` prior to each call to
+    /// `start_send`.
     ///
-    ///  - `Poll::Ready(Ok(()))` if the in-progress send has been completed, or there is
-    ///    no send in progress (even if the channel is closed).
-    ///  - `Poll::Ready(Err(err))` if the in-progress send failed because the channel has
-    ///    been closed.
-    ///  - `Poll::Pending` if a send is in progress, but it could not complete now.
+    /// This method returns `Poll::Ready` once the underlying channel is ready to receive a value,
+    /// by reserving a slot in the channel for the item to be sent. If this method returns
+    /// `Poll::Pending`, the current task is registered to be notified (via
+    /// `cx.waker().wake_by_ref()`) when `poll_ready` should be called again.
     ///
-    /// When this method returns `Poll::Pending`, the current task is scheduled
-    /// to receive a wakeup when the message is sent, or when the entire channel
-    /// is closed (but not if just this sender is closed by
-    /// `close_this_sender`). Note that on multiple calls to `poll_send_done`,
-    /// only the `Waker` from the `Context` passed to the most recent call is
-    /// scheduled to receive a wakeup.
-    ///
-    /// If this method returns `Poll::Ready`, then `start_send` is guaranteed to
-    /// not panic.
-    pub fn poll_send_done(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), SendError<T>>> {
-        if !self.is_sending {
-            return Poll::Ready(Ok(()));
-        }
+    /// # Errors
+    /// 
+    /// If the channel is closed, an error will be returned.  This is a permanent state.
+    /// 
+    /// # Panics
+    /// 
+    /// If a previous call to `poll_reserve` returned `Poll::Ready(Ok(()))`, and `poll_reserve` is
+    /// called again immediately after without a call to `start_send` first, the second call to
+    /// `poll_reserve` will panic.
+    pub fn poll_reserve(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), PollSendError<T>>> {
+        let (result, next_state) = match self.take_state() {
+            State::Idle(sender) => {
+                self.acquire.set(make_acquire_future(Some(sender)));
+                match self.acquire.poll(cx) {
+                    // Channel has capacity.
+                    Poll::Ready(Ok(permit)) => (Poll::Ready(Ok(())), State::ReadyToSend(permit)),
+                    // Channel is closed.
+                    Poll::Ready(Err(e)) => (Poll::Ready(Err(e)), State::Closed),
+                    // Channel doesn't have capacity yet, so we need to wait.
+                    Poll::Pending => (Poll::Pending, State::Acquiring),
+                }
+            },
+            State::Acquiring => match ready!(self.acquire.poll(cx)) {
+                // Channel has capacity.
+                Ok(permit) => (Poll::Ready(Ok(())), State::ReadyToSend(permit)),
+                // Channel is closed.
+                Err(e) => (Poll::Ready(Err(e)), State::Closed), 
+            },
+            // We're closed, either by choice or because the underlying sender was closed.
+            State::Closed => return Poll::Ready(Err(PollSendError(None))),
+            State::ReadyToSend(_) => panic!("called `poll_ready` when `start_send` was expected"),
+        };
 
-        let result = self.inner.poll(cx);
-        if result.is_ready() {
-            self.is_sending = false;
-        }
-        if let Poll::Ready(Err(_)) = &result {
-            self.sender = None;
-        }
+        self.state = next_state;
         result
     }
 
-    /// Check whether the channel is ready to send more messages now.
-    ///
-    /// If this method returns `true`, then `start_send` is guaranteed to not
-    /// panic.
-    ///
-    /// If the channel is closed, this method returns `true`.
-    pub fn is_ready(&self) -> bool {
-        !self.is_sending
+    /// Sends an item to the channel.
+    /// 
+    /// Before calling `start_send`, `poll_reserve` must be called with a successful return
+    /// value of `Poll::Ready(Ok(()))`.
+    /// 
+    /// # Errors
+    /// 
+    /// If the channel is closed, an error will be returned.  This is a permanent state.
+    /// 
+    /// # Panics
+    /// 
+    /// If `poll_reserve` was not successfully called prior to calling `start_send`, then this method
+    /// will panic.
+    pub fn start_send(&mut self, value: T) -> Result<(), PollSendError<T>> {
+        let (result, next_state) = match self.take_state() {
+            State::Idle(_) | State::Acquiring => panic!("`start_send` called without first calling `poll_ready`"),
+            State::ReadyToSend(permit) => (Ok(()), State::Idle(permit.send(value))),
+            State::Closed => (Err(PollSendError(Some(value))), State::Closed),
+        };
+
+        // Handle deferred closing if `close_this_sender` was called between `poll_reserve` and `start_send`.
+        self.state = if self.sender.is_some() { 
+            next_state
+        } else {
+            State::Closed
+        };
+        result
     }
 
     /// Check whether the channel has been closed.
     pub fn is_closed(&self) -> bool {
-        match &self.sender {
-            Some(sender) => sender.is_closed(),
-            None => true,
-        }
+        matches!(self.state, State::Closed)
     }
 
-    /// Clone the underlying `Sender`.
+    /// Clones the underlying `Sender`.
     ///
-    /// If this method returns `None`, then the channel is closed. (But it is
-    /// not guaranteed to return `None` if the channel is closed.)
+    /// If the channel is closed, `None` is returned.
     pub fn clone_inner(&self) -> Option<Sender<T>> {
-        self.sender.as_ref().map(|sender| (&**sender).clone())
+        self.sender.clone()
     }
 
-    /// Access the underlying `Sender`.
-    ///
-    /// If this method returns `None`, then the channel is closed. (But it is
-    /// not guaranteed to return `None` if the channel is closed.)
+    /// Gets a reference to the underlying `Sender`.
+    /// 
+    /// If the channel is closed, `None` is returned.
     pub fn inner_ref(&self) -> Option<&Sender<T>> {
-        self.sender.as_deref()
+        self.sender.as_ref()
     }
 
-    // This operation is supported because it is required by the Sink trait.
     /// Close this sender. No more messages can be sent from this sender.
     ///
-    /// Note that this only closes the channel from the view-point of this
-    /// sender. The channel remains open until all senders have gone away, or
-    /// until the [`Receiver`] closes the channel.
+    /// Note that this only closes the channel from the view-point of this sender. The channel
+    /// remains open until all senders have gone away, or until the [`Receiver`] closes the channel.
     ///
-    /// If there is a send in progress when this method is called, that send is
-    /// unaffected by this operation, and `poll_send_done` can still be called
-    /// to complete that send.
+    /// If a slot was previously reserved by calling `poll_reserve`, then a final call can be made
+    /// to `start_send` in order to consume the reserved slot.  After that, no further sends will be
+    /// possible.  If you do not intend to send another item, you can release the reserved slot back
+    /// to the underlying sender by calling [`abort_send`].
     ///
     /// [`Receiver`]: tokio::sync::mpsc::Receiver
     pub fn close_this_sender(&mut self) {
+        // Mark ourselves officially closed by dropping our main sender.
         self.sender = None;
+
+        // We don't want to abort an in-progress send since the caller might still want to complete
+        // it.  If that's the case, we'll defer updating our state until `start_send` is called, but
+        // otherwise, we adjust our state here.
+        if let State::Idle(_) = self.state {
+            self.state = State::Closed;
+        }
     }
 
-    /// Abort the current in-progress send, if any.
+    /// Aborts the current in-progress send, if any.
     ///
-    /// Returns `true` if a send was aborted.
+    /// Returns `true` if a send was aborted.  If the sender was closed prior to calling
+    /// `abort_send`, then the sender will remain in the closed state, otherwise the sender will be
+    /// ready to attempt another send.
     pub fn abort_send(&mut self) -> bool {
-        if self.is_sending {
-            self.inner.set(make_future(None));
-            self.is_sending = false;
-            true
-        } else {
-            false
-        }
+        let (result, next_state) = match self.take_state() {
+            State::Acquiring => {
+                // Resetting the future drops the in-flight one.
+                self.acquire.set(make_acquire_future(None));
+                let state = match self.sender.clone() {
+                    Some(sender) => State::Idle(sender),
+                    None => State::Closed,
+                };
+                (true, state)
+            },
+            State::ReadyToSend(permit) => {
+                let state = if self.sender.is_some() {
+                    State::Idle(permit.release())
+                } else {
+                    State::Closed
+                };
+                (true, state)
+            },
+            s => (false, s),
+        };
+
+        self.state = next_state;
+        result
     }
 }
 
 impl<T> Clone for PollSender<T> {
-    /// Clones this `PollSender`. The resulting clone will not have any
-    /// in-progress send operations, even if the current `PollSender` does.
+    /// Clones this `PollSender`.
+    /// 
+    /// The resulting `PollSender` will have an initial state identical to calling `PollSender::new`.
     fn clone(&self) -> PollSender<T> {
+        let (sender, state) = match self.sender.clone() {
+            Some(sender) => (Some(sender.clone()), State::Idle(sender)),
+            None => (None, State::Closed),
+        };
+        
         Self {
-            sender: self.sender.clone(),
-            is_sending: false,
-            inner: ReusableBoxFuture::new(async { unreachable!() }),
+            sender,
+            state,
+            acquire: ReusableBoxFuture::new(async { unreachable!() }),
         }
     }
 }
 
 impl<T: Send + 'static> Sink<T> for PollSender<T> {
-    type Error = SendError<T>;
+    type Error = PollSendError<T>;
 
-    /// This is equivalent to calling [`poll_send_done`].
-    ///
-    /// [`poll_send_done`]: PollSender::poll_send_done
     fn poll_ready(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Pin::into_inner(self).poll_send_done(cx)
+        Pin::into_inner(self).poll_reserve(cx)
     }
 
-    /// This is equivalent to calling [`poll_send_done`].
-    ///
-    /// [`poll_send_done`]: PollSender::poll_send_done
-    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Pin::into_inner(self).poll_send_done(cx)
+    fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
     }
 
-    /// This is equivalent to calling [`start_send`].
-    ///
-    /// [`start_send`]: PollSender::start_send
     fn start_send(self: Pin<&mut Self>, item: T) -> Result<(), Self::Error> {
         Pin::into_inner(self).start_send(item)
     }
 
-    /// This method will first flush the `PollSender`, and then close it by
-    /// calling [`close_this_sender`].
-    ///
-    /// If a send fails while flushing because the [`Receiver`] has gone away,
-    /// then this function returns an error. The channel is still successfully
-    /// closed in this situation.
-    ///
-    /// [`close_this_sender`]: PollSender::close_this_sender
-    /// [`Receiver`]: tokio::sync::mpsc::Receiver
-    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        ready!(self.as_mut().poll_flush(cx))?;
-
+    fn poll_close(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
         Pin::into_inner(self).close_this_sender();
         Poll::Ready(Ok(()))
     }